Добавил отдельный файл по монге (класс)
