+++ requirements.txt
@@ -1,3 +1,3 @@
 python-telegram-bot==13.15
 pymongo==4.6.1
-dnspython==2.4.2
+
