+++ source/bot.py
@@ -6,7 +6,8 @@ from telegram.ext import (
 )
 
 from telegram import ReplyKeyboardMarkup, ReplyKeyboardRemove, ParseMode
-from pymongo import MongoClient
+
+from mongo_db import Mongo
 
 
 class Bot:
@@ -15,8 +16,7 @@ class Bot:
     def __init__(self, token):
         self.mybot = Updater(token, use_context=True)
         dp = self.mybot.dispatcher
-        self.client = MongoClient('mongodb://localhost:27017/') #-- Добавление подключения к локальной базе
-        self.db = self.client['Users_LP31_Pars'] #-- название БД
+        self.mongo = Mongo()
         
         dp.add_handler(CommandHandler('start', self.hello_user))
    
@@ -53,7 +53,7 @@ class Bot:
 
     """Функция приветсвует пользователя по имени"""
     def hello_user(self, update, context):
-        self.get_or_create_user_in_db(update.effective_user, update.effective_chat.id)
+        self.mongo.get_or_create_user_in_db(update.effective_user, update.effective_chat.id)
         self.name = update.message.from_user.first_name
         update.message.reply_text(f'Привет, {self.name}! Очень рад.', reply_markup=self.main_keyboard())
         update.message.reply_text(f'''В этом боте ты можешь запросить свежие объявления
@@ -121,18 +121,7 @@ class Bot:
     """
 
 
-    def get_or_create_user_in_db(self, effective_user, chat_id):
-        user = self.db.users.find_one({"user_id": effective_user.id})
-        if not user:
-            user = {
-                "user_id": effective_user.id,
-                "first_name": effective_user.first_name,
-                "last_name": effective_user.last_name,
-                "username": effective_user.username,
-                "chat_id": chat_id
-            }
-            self.db.users.insert_one(user)
-        return user
+
